move sphinx extension to extensions
